Cleaning up multiprocessing.
