+++ dynamo_consistency/threads.py
@@ -1,0 +1,280 @@
+"""
+Module that contains all of the threading functions needed to create a
+:py:class:`dynamo_consistency.datatypes.DirectoryInfo` object.
+
+:author: Daniel Abercrombie <dabercro@mit.edu>
+"""
+
+
+import os
+import time
+import logging
+import multiprocessing
+import threading
+import collections
+
+from Queue import Empty
+
+from . import config
+from .datatypes import DirectoryInfo
+
+
+LOG = logging.getLogger(__name__)
+
+
+def create_dirinfo( # pylint: disable=too-many-locals, too-many-branches, too-many-statements, too-complex
+        location, first_dir, filler, object_params=None, callback=None):
+    """ Create the directory information
+
+    :param str location: This is the beginning of the path where we will find ``first_dir``.
+                         For example, to find the first directory ``mc``, we also have to
+                         say where it is. For using CMS LFNs, location would be
+                         ``/store`` (where ``mc`` is inside).
+                         This is a path.
+    :param str first_dir: The name of the first directory that is inside the path of ``location``.
+                          This should not be a path,
+                          but the name of the directory to list recursively.
+    :param filler: This is either a function that lists the directory contents given just a path
+                   of ``os.path.join(location, first_dir)``, or it is a constructor that
+                   does the same thing with a member function called ``list``.
+                   If ``filler`` is an object constructor, the parameters for the object
+                   creation must be passed through the parameter ``object_params``.
+                   Both listings must return the following tuple:
+
+                     - A bool saying whether the listing was successful or not
+                     - A list of tuples of sub-directories and their mod times
+                     - A list of tuples files inside, their size, and their mode times
+
+    :type filler: function or constructor
+    :param list object_params: This only needs to be set when filler is an object constructor.
+                               Each element in the list is a tuple of arguments to pass
+                               to the constructor.
+    :param function callback: A function that is called every time master thread has finished
+                              checking the child threads.
+                              This can happen very many times at large sites.
+                              The function is called with the main DirectoryTree as its argument
+    :returns: A :py:class:`DirectoryInfo` object containing everything the directory listings from
+              ``os.path.join(location, first_dir)`` with name ``first_dir``.
+    :rtype: DirectoryInfo
+    """
+
+    LOG.debug('Called create_dirinfo(%s, %s, %s, %s)',
+              location, first_dir, filler, object_params)
+
+    # Determine the number of threads
+    if object_params is not None:
+        n_threads = len(object_params)
+    else:
+        n_threads = config.config_dict()['NumThreads'] or multiprocessing.cpu_count()
+
+    # First directory is location + first_dir
+    starting_dir = os.path.join(location, first_dir)
+    LOG.info('Listing directory %s with %i threads', starting_dir, n_threads)
+
+    # Initialize queue and connection lists
+    out_queue = multiprocessing.Queue()
+
+    # We want to add and remove directories from the same side
+    in_lock = threading.Lock()
+    in_deque = collections.deque()
+
+    master_conns = []
+    slave_conns = []
+    send_to_master = []
+
+    for _ in xrange(n_threads):
+        con1, con2 = multiprocessing.Pipe()
+        master_conns.append(con1)
+        slave_conns.append(con2)
+        send_to_master.append(multiprocessing.Queue())
+
+    # Put in the first element for the queue
+    # They go like, (full path of the next listing to do,
+    #                name of sub-node to place the listing (blank for first level),
+    #                list of previous directories, list of previous files (for retries),
+    #                list of queue numbers that have failed so far)
+    in_deque.append((starting_dir, '', [], [], []))
+
+    def run_queue(i_queue):
+        """
+        Runs over one of the queues.
+        When the queue is finished, it checks back with the master for permission to stop.
+
+        :param int i_queue: The number of the thread
+        """
+
+        thread_log = logging.getLogger('%s--thread%i' % (__name__, i_queue))
+
+        thread_log.debug('Running queue: %i', i_queue)
+        running = True
+
+        # Get the queue and connection for this thread
+        conn = slave_conns[i_queue]
+
+        if object_params:
+            # Create the object with the parameters here
+            params = object_params[i_queue]
+            thread_log.debug('Params for this object: %s', params)
+            thread_object = filler(*params)
+            filler_func = thread_object.list
+        else:
+            # Otherwise, use the filler function directly passed
+            filler_func = filler
+
+        while running:
+            try:
+                in_lock.acquire()
+                location, name, prev_dirs, prev_files, failed_list = in_deque.pop()
+                if i_queue in failed_list:
+                    thread_log.warning('Got previously failed call, putting back and sleeping')
+                    in_deque.append((location, name, prev_dirs, prev_files, failed_list))
+
+                in_lock.release()
+
+                if i_queue in failed_list:
+                    time.sleep(10)
+
+                thread_log.debug('Getting directory with (%s, %s, %s)',
+                                 location, name, failed_list)
+
+                # Call filler
+                full_path = os.path.join(location, name)
+                thread_log.debug('Full path is %s', full_path)
+                okay, directories, files = filler_func(full_path)
+                thread_log.debug('Got from filler: Good? %s, %i directories, %i files',
+                                 okay, len(directories), len(files))
+
+                # If not okay, add _unlisted_ flag
+                if not okay:
+
+                    directories = list(set(directories + prev_dirs))
+                    files = list(set(files + prev_files))
+
+                    thread_log.debug('Full dirs, and files: %s, %s', directories, files)
+
+                    thread_log.error('Giving up directory %s', full_path)
+                    # _unlisted_ is used as a flag to tell our comparer something went wrong
+                    files.append(('_unlisted_', 0, 0))
+
+                # Send results to master queue
+                out_queue.put((name, directories, files, len(failed_list)))
+
+                # Add each directory into some input queue
+                for directory, _ in directories:
+                    joined_name = os.path.join(name, directory)
+                    in_lock.acquire()
+                    in_deque.append((location, joined_name, [], [], []))
+                    in_lock.release()
+
+                # Tell master that a job finished,
+                # so it can build the final object
+                send_to_master[i_queue].put(('O', time.time()))
+
+                thread_log.debug('Finished one job with (%s, %s)', location, name)
+            except IndexError:
+                in_lock.release()
+
+                # Report empty
+                thread_log.debug('Worker finished input queue')
+                send_to_master[i_queue].put(('A', 0))
+                #conn.send('All_Job')
+
+                # Check for main process
+                message = conn.recv()
+                thread_log.debug('Message from master: %s', message)
+                # If permission, close
+                if message == 'Close':
+                    conn.close()
+                    running = False
+                else:
+                    thread_log.debug('Worker going back to check queue')
+
+    # Spawn processes to run on this run_queue function
+    processes = []
+
+    for i_queue in range(n_threads):
+        process = multiprocessing.Process(target=run_queue, args=(i_queue,))
+        process.start()
+        processes.append(process)
+
+    # Build the DirectoryInfo
+    building = True
+    dir_info = DirectoryInfo(first_dir)
+
+    while building:
+        try:
+            # Get the info from the queue
+            name, directories, files, _ = out_queue.get(True, 1)
+
+            # Create the nodes and files
+            built = dir_info.get_node(name)
+            built.add_files(files)
+
+            # Set correct node mtime for directories
+            for directory, mtime in directories:
+                built.get_node(directory).mtime = mtime
+
+        except Empty:
+            # When empty, check on the status of the workers
+            LOG.debug('Empty queue for building.')
+            LOG.info('Number of files so far built: %8i  nodes: %8i',
+                     dir_info.get_num_files(), dir_info.count_nodes())
+
+            # Process the dir_info with some callback
+            if callback:
+                callback(dir_info)
+
+            # Ends only if all threads are done at the beginning of this check
+            threads_done = 0
+            for conn in master_conns:
+                LOG.debug('Waiting for thread %i', master_conns.index(conn))
+                message, timestamp = send_to_master[master_conns.index(conn)].get(True)
+
+                LOG.debug('Recieved message %s', message)
+                # Count the number of threads saying their finished at the beginning
+                if message == 'A':
+                    threads_done += 1
+                    LOG.info('Threads saying done: %i', threads_done)
+                    # Send back to work, just in case not all threads are done
+                    conn.send('Work')
+                elif message == 'O':
+                    # This thread wasn't finished at the beginning, so threads_done
+                    # will not reach n_threads if the master reaches this point in the code
+                    LOG.debug('Found one job, about to cycle')
+                    now = time.time()
+                    cycle = True
+                    while cycle:
+                        # Cycle through timestamps so that we do not have a backlog
+                        try:
+                            message, timestamp = \
+                                send_to_master[master_conns.index(conn)].get(True, 1)
+                            if message == 'A':
+                                LOG.debug('Found end to pipe.')
+                                conn.send('Work')
+                                cycle = False
+                            elif timestamp > now:
+                                cycle = False
+                        except Empty:
+                            cycle = False
+
+                else:
+                    LOG.error('Weird message from pipe')
+
+            # Check if all the threads were finished
+            if threads_done == n_threads:
+                LOG.debug('Done building')
+                # Break out of loop of checking
+                building = False
+
+    LOG.debug('Closing all connections')
+    # Tell connections to close
+    for conn in master_conns:
+        conn.send('Close')
+        conn.close()
+
+    LOG.debug('Waiting for processes')
+    # Wait for processes to join
+    for proc in processes:
+        proc.join()
+
+    return dir_info
