+++ test/test_datatypes.py
@@ -33,6 +33,7 @@ from base import TestBase
 from base import TMP_DIR
 
 from dynamo_consistency import datatypes
+from dynamo_consistency import threads
 
 LOG = logging.getLogger(__name__)
 
@@ -136,12 +137,12 @@ class TestConsistentTrees(TestBase):
             LOG.info('In callback: tree has %i files', tree.get_num_files())
             called['check'] = True
 
-        datatypes.create_dirinfo('', 'mc', my_ls, callback=call)
+        threads.create_dirinfo('', 'mc', my_ls, callback=call)
         self.assertTrue(called['check'])
 
     def test_ls_vs_list(self):
 
-        dirinfos = [datatypes.create_dirinfo('', subdir, my_ls) \
+        dirinfos = [threads.create_dirinfo('', subdir, my_ls) \
                         for subdir in ['mc', 'data']]
 
         master_dirinfo = datatypes.DirectoryInfo('/store', directories=dirinfos)
@@ -154,7 +155,7 @@ class TestConsistentTrees(TestBase):
 
         os.makedirs(os.path.join(TMP_DIR, empty_dir))
 
-        dirinfos = [datatypes.create_dirinfo('', subdir, my_ls) \
+        dirinfos = [threads.create_dirinfo('', subdir, my_ls) \
                         for subdir in ['mc', 'data']]
 
         master_dirinfo = datatypes.DirectoryInfo('/store', directories=dirinfos)
@@ -201,7 +202,7 @@ class TestInconsistentTrees(TestBase):
             out.close()
 
         self.listing = datatypes.DirectoryInfo(
-            '/store', directories=[datatypes.create_dirinfo('', subdir, my_ls)
+            '/store', directories=[threads.create_dirinfo('', subdir, my_ls)
                                    for subdir in ['mc', 'data']])
 
         self.tree.setup_hash()
@@ -242,7 +243,7 @@ class TestInconsistentTrees(TestBase):
         os.utime(path, (1000000000, 1000000000))
 
         listing = datatypes.DirectoryInfo('/store',
-                                          directories=[datatypes.create_dirinfo('', subdir, my_ls)
+                                          directories=[threads.create_dirinfo('', subdir, my_ls)
                                                        for subdir in ['mc', 'data']])
 
         LOG.info('='*40)
